Migration away from deprecated generate RPC
1) Remove deprecated BitcoinClient#generateBlock and 
   BitcoinClient#generateBlocks
2) Add BitcoinExtendedClient#generateBlocks that wraps generateToAddress
   to ease the transition
3) Add regTestMiningAddress (using a Bitcoin Core wallet address) to
   BitcoinExtendedClient so mined coins are captured and usable
4) Use BitcoinExtendedClient#generateBlocks in the various RegTest
   integration tests and fixtures in this repository.

I’m assuming that Bitcoin Core is trying to migrate RegTest tools away
from reliance on Bitcoin Core wallet addresses. So I’m imagining a
second step in this migration, which is to generate a key in the client
and have regTestMiningAddress be a client key/address. This will require
changes in the RegTest integration test fixtures in both ConsensusJ and
OmniJ. It may also be possible to use generateToAddress to more
directly/quickly get funds in addresses used in functional tests.