JSON-RPC clients: prefer SSLContext over SSLSocketFactory
As a step towards migration to the java.net.http JSON-RPC client, we
need to migrate our SSL configuration from using SSLSocketFactory to
using SSLContext.  (SSLContext can return an SSLSocketFactory, but
java.net.http.HttpClient needs to be configured with an SSLContext)

Provide deprecated constructors for AbstractRpcClient subclasses that
had used SSLSocketFactory, but migrate the CLI-related classes in
a breaking way.