BitcoinClient: Add importDescriptor subset functionality
This is a step towards a regTest client that can mine using a descriptor
wallet. Nix Darwin builds of Bitcoin Core don't include the legacy wallet,
so our tests can't run against a Nix-built bitcoind on Darwin.

So a short-term goal is to do the minimum necessary to run regTest
against a bitciond instance without legacy wallet support.

This PR enables regTest mining to work with descriptor wallets, but
the useLegacyWallet flag is hardcoded to true because even with the
mining support there are issues in some of the tests so they
won't work with descriptor wallets. see the comment at line 130
in BitcoinExtendedClient.java.